pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sift-kg"
version = "0.6.0"
description = "Zero-config document-to-knowledge-graph pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Juan Ceresa", email = "jcere@umich.edu"}
]
keywords = ["knowledge-graph", "nlp", "llm", "document-processing", "entity-extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.1",
    "pdfplumber>=0.10.0",
    "kreuzberg>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "networkx>=3.2",
    "litellm>=1.0.0",
    "unidecode>=1.3.0",
    "semhash>=0.4.0",
    "inflect>=7.0.0",
    "pyvis>=0.3.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.0.0",
    "scikit-learn>=1.3.0",
]
ocr = [
    "google-cloud-vision>=3.4.0",
    "pymupdf>=1.23.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.9.0",
]
all = [
    "sift-kg[embeddings]",
    "sift-kg[ocr]",
]

[project.scripts]
sift = "sift_kg.cli:app"

[project.urls]
Homepage = "https://github.com/civictable/sift-kg"
Documentation = "https://github.com/civictable/sift-kg#readme"
Repository = "https://github.com/civictable/sift-kg"
Issues = "https://github.com/civictable/sift-kg/issues"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
